Added authenticatoin to views

+++ products/views.py
@@ -5,6 +5,7 @@ from .models import Product
 from .forms import ProductForm
 
 from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -64,8 +65,11 @@ def product_detail(request, product_id):
 
     return render(request, "products/product_detail.html", context)
 
-
+@login_required
 def add_product(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'Only super users can do that')
+        return redirect(reverse('home'))
 
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
@@ -84,7 +88,11 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Only super users can do that')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     if request.method == "POST":
@@ -107,7 +115,11 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Only super users can do that')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
